// src/types/ArtistStrings.interface.ts
export interface ArtistPageSections {
  academic: string;
  experience: string;
  skills: string;
  languages: string;
}

export interface ArtistPageButtons {
  view_cv: string;
  download_cv: string;
  contact: string;
}

export interface ArtistPageStrings {
  sections: ArtistPageSections;
  buttons: ArtistPageButtons;
}